Upgrade Elixir & Erlang + deps.
We cannot use the httpc adapter with Tesla since it does not work for
OTP27 due to an incompatibility in the httpc library. Thus we replaced
it with the recommended adapter (Mint). It's only for testing anyway...

While we're at it, we restructure the test support setup and add the
folder to the elixir compile paths